Moved Active and Maintenance into command group
